Fix MinGW amd64 alloc tag prototypes

+++ MinGW/amd64/include/lib9.h
@@ -186,10 +186,10 @@ extern	void	free(void*);
 extern	ulong	msize(void*);
 extern	void*	calloc(size_t, size_t);
 extern	void*	realloc(void*, size_t);
-extern	void		setmalloctag(void*, ulong);
-extern	void		setrealloctag(void*, ulong);
-extern	ulong	getmalloctag(void*);
-extern	ulong	getrealloctag(void*);
+extern	void		setmalloctag(void*, uintptr);
+extern	void		setrealloctag(void*, uintptr);
+extern	uintptr	getmalloctag(void*);
+extern	uintptr	getrealloctag(void*);
 extern	void*	malloctopoolblock(void*);
 
 #define malloc	kmalloc
